Remove the need for dev cleanup

+++ setup.py
@@ -26,17 +26,32 @@ class CMakeExtension(BaseExtension):
 
     def __init__(self, name, build_dir, *args, **kwargs):
         super().__init__(name, [], *args, **kwargs)
+        self.ext_sfx = importlib.machinery.EXTENSION_SUFFIXES[0]
         self.build_dir = Path(build_dir)
 
+    def ext_list(self):
+        return list(self.build_dir.glob(f"**/*{self.ext_sfx}"))
+
+    def exists(self):
+        return len(self.ext_list()) > 0
+
+    def ext_file_path(self):
+        if not self.exists():
+            return None
+        return self.ext_list()[0]
+
     def prepare(self, builder):
-        if self.build_dir.exists():
+        # Only used for dev purposes
+        if self.build_dir.exists() and (os.environ.get("DEV") is not None):
             for root, dirs, files in os.walk(self.build_dir):
                 for f in (dirs + files):
                     os.chmod(Path(root) / f, 0o777)
             
             shutil.rmtree(self.build_dir)
         
-        self.build_dir.mkdir()
+        self.build_dir.mkdir(exist_ok=True)
+        if self.exists():
+            self.ext_file_path().unlink()
 
     def configure(self, builder):
         subprocess.run( ["cmake", ".."], cwd=self.build_dir)
@@ -45,14 +60,12 @@ class CMakeExtension(BaseExtension):
         subprocess.run( ["cmake", "--build", "."], cwd=self.build_dir)
 
     def install(self, builder):
-        sfx = importlib.machinery.EXTENSION_SUFFIXES[0]
-        self.ext_file_path = next(self.build_dir.glob(f"**/*{sfx}"))
-        if self.ext_file_path.exists():
+        if self.ext_file_path().exists():
             path = Path(builder.get_ext_fullpath(self.name))
             path.parent.mkdir(parents=True)
 
             shutil.copyfile(
-                str(self.ext_file_path),
+                str(self.ext_file_path()),
                 builder.get_ext_fullpath(self.name)
             )
         else:
